Fix an issue cased by `jq` in catalina
The issue details are in https://github.com/stedolan/jq/issues/1992, but this doesn’t seem like something we can fix, or even work around while still using `jq` for this particular functionality (if we can’t use the `test` operator, we can’t filter the list, AFAICT).

This change uses a really simple ruby script to parse the list and get us the same data.