Add tip to halt message in should-publish-library-artifacts
It can happen that a developer pushes a branch for a finished feature
and then opens a PR for it.

In such case, depending on the load on the CircleCI queue, the build
might start before the PR is open, which would result in the S3
publishing halting, as intended.

There's not much we can do from a tooling perspective, short of building
and upload every source code check in, to work around this issue.

This commit adds what is hopefully going to be useful information for
developers confused by this behavior.